+++ Twilio/Tests/Integration/Preview/Understand/Service/ModelBuildTest.php
@@ -43,7 +43,9 @@ class ModelBuildTest extends HolodeckTestCase {
                 "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "unique_name": "unique_name"
+                "unique_name": "unique_name",
+                "build_duration": null,
+                "error_code": null
             }
             '
         ));
@@ -113,11 +115,13 @@ class ModelBuildTest extends HolodeckTestCase {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "date_updated": "2015-07-30T20:00:00Z",
                         "url": "https://preview.twilio.com/understand/Services/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/ModelBuilds/UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        "status": "enqueued",
+                        "status": "failed",
                         "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "date_created": "2015-07-30T20:00:00Z",
                         "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        "unique_name": "unique_name"
+                        "unique_name": "unique_name",
+                        "build_duration": null,
+                        "error_code": 23001
                     }
                 ]
             }
@@ -157,7 +161,9 @@ class ModelBuildTest extends HolodeckTestCase {
                 "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "unique_name": "unique_name"
+                "unique_name": "unique_name",
+                "build_duration": null,
+                "error_code": null
             }
             '
         ));
@@ -191,11 +197,13 @@ class ModelBuildTest extends HolodeckTestCase {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_updated": "2015-07-30T20:00:00Z",
                 "url": "https://preview.twilio.com/understand/Services/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/ModelBuilds/UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "enqueued",
+                "status": "completed",
                 "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "unique_name": "unique_name"
+                "unique_name": "unique_name",
+                "build_duration": 100,
+                "error_code": null
             }
             '
         ));